requirements.txt
Flask~=3.0.0
Flask-SQLAlchemy~=3.1.1
PyMySQL~=1.1.0
cryptography~=41.0.7